Instanciando músicas e podcasts

+++ src/br/com/alura/minhasmusicas/modelos/Audio.java
@@ -4,7 +4,7 @@ public class Audio {
     private String titulo;
     private int totalReproducoes;
     private int totalCurtidas;
-    private double classificacao;
+    private int classificacao;
 
     public String getTitulo() {
         return titulo;
@@ -22,12 +22,12 @@ public class Audio {
         return totalCurtidas;
     }
 
-    public double getClassificacao() {
+    public int getClassificacao() {
         return classificacao;
     }
 
     public void curte() {
-        this.totalCurtidas ++;
+        this.totalCurtidas++;
     }
 
     public void reproduz() {
